+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file. Be aware th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Support for [the Priority plugin](https://www.curseforge.com/wow/addons/betterbags-priority).
+
 ## [1.4.0] 2024-08-16
 ### Added
 - The plugin now clears items from the BoE/WoE categories when equipping them. For now, this works only in Retail.